feat: 클릭하면 장르 나오고 들어가도록(화살표도 바뀌게)/component로
