+++ src/pages/profile-upload/index.tsx
@@ -2,8 +2,9 @@ import { useRef, useState, useMemo } from 'react';
 import Image from 'next/image';
 import styles from '../../styles/UploadPage.module.css';
 import BlankImage from '../../assets/icons/blank-image.svg';
-import UploadVideo from '../../components/video/UploadVideo';
 import { uploadFile } from '../../types/upload';
+import UploadVideo from '../../components/upload/UploadVideo';
+import DanceGenre from '@/components/upload/DanceGenre';
 
 export default function ProfileUpload() {
   const [image, setImage] = useState<uploadFile | null>(null);
@@ -15,6 +16,7 @@ export default function ProfileUpload() {
   const [hashTag, setHashTag] = useState<string>('');
   const [date, setDate] = useState<string>('');
   const [award, setAward] = useState<string>('');
+  const [clicked, isClicked] = useState<boolean>(false);
 
   //image 미리보기
   const fileInputRef = useRef<HTMLInputElement>(null);
@@ -57,44 +59,10 @@ export default function ProfileUpload() {
     );
   }, [image]);
 
-  //genre 선택
-  const genreList = [
-    { name: '락킹', length: 2 },
-    { name: '왁킹', length: 2 },
-    { name: '팝핑', length: 2 },
-    { name: '재즈', length: 2 },
-    { name: '걸스 힙합', length: 5 },
-    { name: '힙합', length: 2 },
-    { name: '텃팅', length: 2 },
-    { name: '디짓', length: 2 },
-    { name: '트월크', length: 3 },
-    { name: '힐댄스', length: 3 },
-    { name: '브레이킹', length: 4 },
-    { name: '크럼프', length: 3 },
-    { name: '하우스', length: 3 },
-    { name: '코레오', length: 3 },
-    { name: '컨템', length: 2 },
-    { name: '보깅', length: 2 },
-    { name: '소울', length: 2 },
-    { name: '프리스타일', length: 5 },
-    { name: '기타', length: 2 },
-  ];
-
-  const clickedBox = genreList.map((data, idx) => {
-    if (data.length == 2) {
-      return (
-        <div className={styles.smallGenreBox} key={idx}>
-          {data.name}
-        </div>
-      );
-    } else {
-      return (
-        <div className={styles.bigGenreBox} key={idx}>
-          {data.name}
-        </div>
-      );
-    }
-  });
+  //Genre 선택
+  const handleOpenBox = () => {
+    isClicked(!clicked);
+  };
 
   return (
     <div className={styles.container}>
@@ -144,15 +112,24 @@ export default function ProfileUpload() {
       </div>
       <div className={styles.box}>
         <div className={styles.text}>나의 장르</div>
-        <input
-          className={`${styles.input} ${styles.genre}`}
-          placeholder="나의 댄스 장르를 선택해주세요"
-          onChange={(e: any) => {
-            setGenre(e.target.value);
-            console.log(genre);
-          }}
-        />
-        <div className={styles.clickedBox}>{clickedBox}</div>
+        {clicked ? (
+          <>
+            <input
+              className={`${styles.input} ${styles.genre} ${styles.after}`}
+              placeholder="나의 댄스 장르를 선택해주세요"
+              onClick={handleOpenBox}
+            />
+            <DanceGenre />
+          </>
+        ) : (
+          <>
+            <input
+              className={`${styles.input} ${styles.genre} ${styles.before}`}
+              placeholder="나의 댄스 장르를 선택해주세요"
+              onClick={handleOpenBox}
+            />
+          </>
+        )}
       </div>
       <div className={styles.box}>
         <div className={styles.text}>나를 소개하는 해시태그</div>
